+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.1.5
+- Fix timeout error on create session
+- Fix transport error on delete session
+
+## v0.1.4
+- Add exception throwing when results truncated
+- lint: add line feed at file end
+
+## v0.1.3
 - Add static auth
 ## v0.1.1
 - Add static code analysis